# public/swagger.yaml
openapi: 3.0.3
info:
  title: Dentista Now API
  description: |
    API completa para gerenciamento de barbearia incluindo agendamentos, barbeiros, clientes, produtos, planos e webhooks.

    ## Autenticação
    A maioria dos endpoints requer autenticação via Bearer Token no header:
    ```
    Authorization: Bearer SEU_TOKEN_AQUI
    ```

    ## Formato de Datas
    - Datas são no formato brasileiro: DD/MM/YYYY ou YYYY-MM-DD
    - Horários são no formato: HH:MM (24h)

    ## Webhooks
    O sistema suporta dois tipos de webhooks:
    1. **Webhook Global**: Configurado na tabela `configuracoes`, dispara para todos os eventos
    2. **Webhooks de Barbeiros**: Webhooks personalizados por profissional

  version: 3.0.0
  contact:
    name: Dentista Now
    url: https://dentista-now.vercel.app
servers:
  - url: https://dentista-now.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação da API
  schemas:
...
    Servico:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        descricao:
          type: string
        preco:
          type: number
        duracao_minutos:
          type: integer
        categoria:
          type: string
        ativo:
          type: boolean
...
paths:
...
  /servicos:
    get:
      tags:
        - Serviços
      summary: Listar serviços ativos
      description: Retorna todos os serviços disponíveis para agendamento
      responses:
        '200':
          description: Lista de serviços
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Servico'

    post:
      tags:
        - Serviços
      summary: Criar novo serviço
      description: Adiciona um novo serviço ao catálogo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - preco
                - duracao_minutos
              properties:
                nome:
                  type: string
                descricao:
                  type: string
                preco:
                  type: number
                duracao_minutos:
                  type: integer
                categoria:
                  type: string
      responses:
        '201':
          description: Serviço criado com sucesso

    put:
      tags:
        - Serviços
      summary: Atualizar serviço
      description: Modifica dados de um serviço existente
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  format: uuid
                nome:
                  type: string
                descricao:
                  type: string
                preco:
                  type: number
                duracao_minutos:
                  type: integer
                categoria:
                  type: string
                ativo:
                  type: boolean
      responses:
        '200':
          description: Serviço atualizado

    delete:
      tags:
        - Serviços
      summary: Desativar serviço
      description: Marca um serviço como inativo (soft delete)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Serviço desativado

  # =====================
  # PRODUTOS
  # =====================
  /produtos/listar:
    get:
      tags:
        - Produtos
      summary: Listar produtos
      description: Retorna todos os produtos do catálogo
      security:
        - BearerAuth: []
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
        - name: categoria
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de produtos
...
  /planos/listar:
    get:
      tags:
        - Planos
      summary: Listar planos de fidelidade
      description: Retorna todos os planos disponíveis
      security:
        - BearerAuth: []
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de planos
...